backend/requirements.txt
# FastAPI and server
fastapi==0.115.5
uvicorn==0.32.1
python-multipart==0.0.20

# Google ADK and GenAI
google-adk==1.18.0
google-genai==1.50.1

# Data validation
pydantic==2.10.3
pydantic-settings==2.6.1

# Environment variables
python-dotenv==1.0.1

# HTTP client
httpx==0.28.1

# Async support
anyio==4.7.0
